chore: move svg files to icons folder
