+++ src/assets/icons/ClearIcon.tsx
