+++ src/assets/icons/CloseIcon.tsx
