+++ src/assets/icons/SearchIcon.tsx
