+++ src/assets/icons/index.ts
@@ -1,0 +1,7 @@
+export { ChevronDownIcon } from './ChevronDownIcon.tsx';
+export { ChevronUpIcon } from './ChevronUpIcon.tsx';
+export { ChevronLeftIcon } from './ChevronLeftIcon.tsx';
+export { ChevronRightIcon } from './ChevronRightIcon.tsx';
+export { SearchIcon } from './SearchIcon.tsx';
+export { CloseIcon } from './CloseIcon.tsx';
+export { ClearIcon } from './ClearIcon.tsx';
