+++ src/features/country-search/CountrySearch.tsx
@@ -1,6 +1,6 @@
 import { type ChangeEvent } from 'react';
 import clsx from 'clsx';
-import { SearchIcon, ClearIcon } from '@/shared/ui/icons';
+import { SearchIcon, ClearIcon } from 'assets/icons';
 import { useCountries } from '@/entities/countries';
 
 interface CountrySearchProps {
