+++ src/features/country-sort/CountrySort.tsx
@@ -1,4 +1,4 @@
-import { ChevronUpIcon, ChevronDownIcon } from '@/shared/ui/icons';
+import { ChevronUpIcon, ChevronDownIcon } from 'assets/icons';
 import { useCountries } from '@/entities/countries';
 import type { SortField } from '@/entities/countries/model/types.ts';
 import clsx from 'clsx';
